// src/utils/pricing/types.ts
export interface SectionCounts {
  sections4x8: number;
  sections4x4: number;
  sections2x8: number;
  sections8x2: number;
  sections2x6: number;
  sections6x2: number;
  sections2x4: number;
  sections4x2: number;
  sections2x2: number;
}

export interface SectionPrices {
  section4x8Price: number;
  section4x4Price: number;
  section2x8Price: number;
  section2x4Price: number;
  section4x2Price: number;
  section2x2Price: number;
}